requirements.txt
aiosqlite
PyJWT
click==7.1.1
fastapi==0.53.1
h11==0.9.0
pydantic==1.4
starlette==0.13.2
uvicorn==0.11.3
websockets==8.1
Jinja2==2.11.2
aiofiles==0.5.0
